+++ .github/workflows/docs-check.yml
@@ -42,8 +42,7 @@ jobs:
       - name: Install dependencies
         run: |
           # python -m pip install --upgrade --user pip
-          pip install --requirement requirements/base.txt --upgrade-strategy only-if-needed --find-links https://download.pytorch.org/whl/cpu/torch_stable.html --quiet
-          pip install --requirement requirements/extra.txt
+          pip install --requirement requirements/devel.txt --upgrade-strategy only-if-needed --find-links https://download.pytorch.org/whl/cpu/torch_stable.html --quiet
           pip install --requirement requirements/docs.txt
           python --version
           pip --version
